Move Arel stuff into plugin
